% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic-report.R
\name{devguide_eic_report}
\alias{devguide_eic_report}
\title{Generate a summary report for incoming Editor-in-Charge of current state of
all open software-review issues.}
\usage{
devguide_eic_report()
}
\value{
(Invisibly) A `data.frame` with one row per issue and some key
statistics.
}
\description{
Generate a summary report for incoming Editor-in-Charge of current state of
all open software-review issues.
}
