% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble-links.R
\name{devguide_links}
\alias{devguide_links}
\title{Table of external links used in the dev guide}
\usage{
devguide_links()
}
\value{
a tibble
}
\description{
Table of external links used in the dev guide
}
