% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{devguide_path}
\alias{devguide_path}
\title{Get local path to `dev_guide` repo}
\usage{
devguide_path(path = ".")
}
\arguments{
\item{path}{Path to start searching for `dev_guide` repo.}
}
\value{
Local path to `dev_guide` repo.
}
\description{
Get local path to `dev_guide` repo
}
