% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerelease.R
\name{devguide_prerelease}
\alias{devguide_prerelease}
\title{Open a pre-release issue on the dev_guide repo}
\usage{
devguide_prerelease(post = TRUE)
}
\arguments{
\item{post}{If `FALSE` return markdown-formatted text of comment that would
be used to open a new issue, otherwise open issue with that text.}
}
\value{
Text of issue comment (invisibly).
}
\description{
Uses the gh cli (https://cli.github.com), which must be installed and
authorised with 'gh auth'. The pre-release issue will include all open issues
assigned to the local git user.
}
