% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thanks.R
\name{devguide_thanks}
\alias{devguide_thanks}
\title{Extract list of people to thank for Dev Guide contributions}
\usage{
devguide_thanks()
}
\description{
Extract list of people to thank for Dev Guide contributions
}
