% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_grooming.R
\name{devguide_urls}
\alias{devguide_urls}
\title{Get and optionally open all URLs used in Dev Guide}
\usage{
devguide_urls()
}
\value{
Information on all URLs which are not okay.
}
\description{
This function can take a long time to run!
}
