% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devguide-package.R
\docType{package}
\name{devguider-package}
\alias{devguider-package}
\alias{_PACKAGE}
\title{devguider: Tools for automating processes to manage rOpenSci's DevGuide}
\description{
Tools for automating processes to manage rOpenSci's DevGuide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci-org/devguider}
  \item Report bugs at \url{https://github.com/ropensci-org/devguider/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{maelle.salmon@yahoo.se} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})

Authors:
\itemize{
  \item Maëlle Salmon \email{maelle.salmon@yahoo.se} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
}

Other contributors:
\itemize{
  \item rOpenSci (https://ropensci.org/) [funder]
}

}
\keyword{internal}
